Move source code in common area
